+++ lib/papa/cli.rb
@@ -1,7 +1,8 @@
 require 'papa/common'
-require 'papa/release'
-require 'papa/hotfix'
-require 'papa/sandbox'
+require 'papa/cli/release'
+require 'papa/cli/hotfix'
+require 'papa/cli/integration'
+require 'papa/cli/sandbox'
 
 module Papa
   class CLI < Thor
@@ -10,6 +11,22 @@ module Papa
 
     desc 'hotfix [COMMAND]', 'Perform actions on hotfix branches'
     subcommand 'hotfix', Hotfix
+
+    desc 'integration [COMMAND]', 'Perform actions on integration branches'
+    subcommand 'integration', Integration
+
+    desc 'deploy', 'Deploy a branch with larga'
+    option :branch, aliases: '-b', required: true
+    option :hostname, aliases: '-h'
+    def deploy
+      require 'papa/deploy'
+      require 'papa/larga'
+
+      branch = options[:branch]
+      hostname = options[:hostname]
+
+      Deploy.new(branch: branch, hostname: hostname).run
+    end
 
     desc 'sandbox [COMMAND]', 'Test out papa in a sandboxed git environment'
     subcommand 'sandbox', Sandbox