+++ lib/papa/cli/integration.rb
@@ -1,0 +1,13 @@
+module Papa
+  class Integration < Thor
+    desc 'start', 'Start an integration branch'
+    option :base_branch, aliases: '-b', required: true
+    def start
+      base_branch = options[:base_branch]
+      
+      require 'papa/common/start'
+      require 'papa/integration/start'
+      Integration::Start.new(base_branch: base_branch).run
+    end
+  end
+end