+++ lib/papa/common/add.rb
@@ -11,6 +11,7 @@ module Papa
         queue.add Git.fetch(remote: 'origin')
         queue.add Git.checkout(branch_name: @build_branch)
         queue.add Git.checkout(branch_name: branch)
+        queue.add Git.hard_reset(remote: 'origin', branch_name: branch)
         queue.add Git.rebase(base_branch_name: @build_branch)
         queue.add Git.force_push(remote: 'origin', branch_name: branch)
         queue.add Git.checkout(branch_name: @build_branch)